+++ src/apis/index.ts
@@ -1,10 +1,12 @@
 /* tslint:disable */
 /* eslint-disable */
 export * from "./APIKeysApi";
+export * from "./BillingApi";
 export * from "./CatalogApi";
 export * from "./ComponentSessionsApi";
 export * from "./HostedSessionsApi";
 export * from "./MembershipsApi";
+export * from "./PatientsApi";
 export * from "./PlatformOrdersApi";
 export * from "./PlatformWebhooksApi";
 export * from "./PlatformsApi";