+++ src/models/ListOrdersResponseDataInner.ts
@@ -96,18 +96,6 @@ export interface ListOrdersResponseDataInner {
   externalOrderId: string;
   /**
    *
-   * @type {boolean}
-   * @memberof ListOrdersResponseDataInner
-   */
-  externalSubmissionAttempted: boolean;
-  /**
-   *
-   * @type {string}
-   * @memberof ListOrdersResponseDataInner
-   */
-  externalSubmissionBlockedReason: string | null;
-  /**
-   *
    * @type {string}
    * @memberof ListOrdersResponseDataInner
    */
@@ -280,16 +268,6 @@ export function instanceOfListOrdersResponseDataInner(
   if (!("directions" in value) || value["directions"] === undefined) return false;
   if (!("dosageForm" in value) || value["dosageForm"] === undefined) return false;
   if (!("externalOrderId" in value) || value["externalOrderId"] === undefined) return false;
-  if (
-    !("externalSubmissionAttempted" in value) ||
-    value["externalSubmissionAttempted"] === undefined
-  )
-    return false;
-  if (
-    !("externalSubmissionBlockedReason" in value) ||
-    value["externalSubmissionBlockedReason"] === undefined
-  )
-    return false;
   if (!("id" in value) || value["id"] === undefined) return false;
   if (!("livemode" in value) || value["livemode"] === undefined) return false;
   if (!("medicationName" in value) || value["medicationName"] === undefined) return false;
@@ -337,8 +315,6 @@ export function ListOrdersResponseDataInnerFromJSONTyped(
     directions: json["directions"],
     dosageForm: json["dosageForm"],
     externalOrderId: json["externalOrderId"],
-    externalSubmissionAttempted: json["externalSubmissionAttempted"],
-    externalSubmissionBlockedReason: json["externalSubmissionBlockedReason"],
     id: json["id"],
     livemode: json["livemode"],
     medicationName: json["medicationName"],
@@ -387,8 +363,6 @@ export function ListOrdersResponseDataInnerToJSONTyped(
     directions: value["directions"],
     dosageForm: value["dosageForm"],
     externalOrderId: value["externalOrderId"],
-    externalSubmissionAttempted: value["externalSubmissionAttempted"],
-    externalSubmissionBlockedReason: value["externalSubmissionBlockedReason"],
     id: value["id"],
     livemode: value["livemode"],
     medicationName: value["medicationName"],